Reorganize the oidc package
